handle large block sizes apparently not handled

+++ application-src/amgtar.c
@@ -1839,7 +1839,7 @@ amgtar_restore(
 		fprintf(dar_file,"DAR 0:-1\n");
 		g_debug("full dar: 0:-1");
 	    } else {
-		int   previous_block = -1;
+		off_t   previous_block = -1;
 
 		while (fgets(line, 32768, recover_state_file) != NULL) {
 		    off_t     block_no = g_ascii_strtoull(line, NULL, 0);
